+++ src/cattrs/preconf/cbor2.py
@@ -1,12 +1,13 @@
 """Preconfigured converters for cbor2."""
-from datetime import datetime, timezone
-from typing import Any, Type, TypeVar
+from datetime import date, datetime, timezone
+from typing import Any, Type, TypeVar, Union
 
 from cbor2 import dumps, loads
 
 from cattrs._compat import AbstractSet
 
 from ..converters import BaseConverter, Converter
+from ..strategies import configure_union_passthrough
 
 T = TypeVar("T")
 
@@ -30,6 +31,9 @@ def configure_converter(converter: BaseConverter):
     converter.register_structure_hook(
         datetime, lambda v, _: datetime.fromtimestamp(v, timezone.utc)
     )
+    converter.register_unstructure_hook(date, lambda v: v.isoformat())
+    converter.register_structure_hook(date, lambda v, _: date.fromisoformat(v))
+    configure_union_passthrough(Union[str, bool, int, float, None, bytes], converter)
 
 
 def make_converter(*args: Any, **kwargs: Any) -> Cbor2Converter: