+++ src/cattrs/preconf/pyyaml.py
@@ -1,15 +1,22 @@
 """Preconfigured converters for pyyaml."""
-from datetime import datetime
-from typing import Any, Type, TypeVar
+from datetime import date, datetime
+from typing import Any, Type, TypeVar, Union
 
 from yaml import safe_dump, safe_load
 
 from cattrs._compat import FrozenSetSubscriptable
 
 from ..converters import BaseConverter, Converter
+from ..strategies import configure_union_passthrough
 from . import validate_datetime
 
 T = TypeVar("T")
+
+
+def validate_date(v, _):
+    if not isinstance(v, date):
+        raise ValueError(f"Expected date, got {v}")
+    return v
 
 
 class PyyamlConverter(Converter):
@@ -26,11 +33,20 @@ def configure_converter(converter: BaseConverter):
 
     * frozensets are serialized as lists
     * string enums are converted into strings explicitly
+    * datetimes and dates are validated
     """
     converter.register_unstructure_hook(
         str, lambda v: v if v.__class__ is str else v.value
     )
+
+    # datetime inherits from date, so identity unstructure hook used
+    # here to prevent the date unstructure hook running.
+    converter.register_unstructure_hook(datetime, lambda v: v)
     converter.register_structure_hook(datetime, validate_datetime)
+    converter.register_structure_hook(date, validate_date)
+    configure_union_passthrough(
+        Union[str, bool, int, float, None, bytes, datetime, date], converter
+    )
 
 
 def make_converter(*args: Any, **kwargs: Any) -> PyyamlConverter: