+++ src/cattrs/preconf/ujson.py
@@ -1,13 +1,14 @@
 """Preconfigured converters for ujson."""
 from base64 import b85decode, b85encode
-from datetime import datetime
-from typing import Any, AnyStr, Type, TypeVar
+from datetime import date, datetime
+from typing import Any, AnyStr, Type, TypeVar, Union
 
 from ujson import dumps, loads
 
 from cattrs._compat import AbstractSet
 
 from ..converters import BaseConverter, Converter
+from ..strategies import configure_union_passthrough
 
 T = TypeVar("T")
 
@@ -35,6 +36,9 @@ def configure_converter(converter: BaseConverter):
 
     converter.register_unstructure_hook(datetime, lambda v: v.isoformat())
     converter.register_structure_hook(datetime, lambda v, _: datetime.fromisoformat(v))
+    converter.register_unstructure_hook(date, lambda v: v.isoformat())
+    converter.register_structure_hook(date, lambda v, _: date.fromisoformat(v))
+    configure_union_passthrough(Union[str, bool, int, float, None], converter)
 
 
 def make_converter(*args: Any, **kwargs: Any) -> UjsonConverter: