+++ tests/typeddicts.py
@@ -59,7 +59,9 @@ def int_attributes(
 def datetime_attributes(
     draw: DrawFn, total: bool = True, not_required: bool = False
 ) -> Tuple[datetime, SearchStrategy, SearchStrategy]:
-    success_strat = datetimes().map(lambda dt: dt.replace(microsecond=0))
+    success_strat = datetimes(
+        min_value=datetime(1970, 1, 1), max_value=datetime(2038, 1, 1)
+    ).map(lambda dt: dt.replace(microsecond=0))
     type = datetime
     strat = success_strat if total else success_strat | just(NOTHING)
     if not_required and draw(booleans()):